requirements.txt

prometheus-client==0.13.1
pyyaml==6.0
schema==0.7.5
websockets==10.4
structlog==22.1.0
requests==2.28.1
